Remove image filter and move to multimodal
