+++ nemo_curator/stages/image/filters/__init__.py
@@ -13,13 +13,9 @@
 # limitations under the License.
 
 from .aesthetic_filter import ImageAestheticFilterStage
-from .blur_filter import ImageBlurFilterStage
 from .nsfw_filter import ImageNSFWFilterStage
-from .qrcode_filter import ImageQRCodeFilterStage
 
 __all__ = [
     "ImageAestheticFilterStage",
-    "ImageBlurFilterStage",
     "ImageNSFWFilterStage",
-    "ImageQRCodeFilterStage",
 ]
